+++ src/routes/translation.ts
@@ -3,7 +3,6 @@ import { Readable } from 'node:stream';
 import { Request, Response, NextFunction, Router } from 'express';
 import { parse, stringify } from 'csv';
 import multer from 'multer';
-import { pick } from 'lodash';
 
 import { logger } from '../utils/logger';
 import { UnknownException } from '../exceptions/unknown.exception';
@@ -11,11 +10,10 @@ import { BadRequestException } from '../exceptions/bad-request.exception';
 import { Dataset } from '../entities/dataset/dataset';
 import { DatasetDTO } from '../dtos/dataset-dto';
 import { TranslationDTO } from '../dtos/translations-dto';
-import { translatableMetadataKeys } from '../types/translatable-metadata';
-import { RelatedLink } from '../dtos/related-link-dto';
 import { EventLog } from '../entities/event-log';
 
 import { loadDataset } from './dataset';
+import { collectTranslations } from '../utils/collect-translations';
 
 export const translationRouter = Router();
 
@@ -23,49 +21,6 @@ const upload = multer({ storage: multer.memoryStorage() });
 
 // imported translation filename can be constant as we overwrite each time it's imported
 const TRANSLATION_FILENAME = 'translation-import.csv';
-
-const collectTranslations = (dataset: Dataset, includeIds = false): TranslationDTO[] => {
-  const revision = dataset.draftRevision!;
-  const metadataEN = revision.metadata?.find((meta) => meta.language.includes('en'));
-  const metadataCY = revision.metadata?.find((meta) => meta.language.includes('cy'));
-
-  // ignore roundingDescription if rounding isn't applied
-  const metadataKeys = translatableMetadataKeys.filter((key) => {
-    return revision.roundingApplied === true ? true : key !== 'roundingDescription';
-  });
-
-  const translations: TranslationDTO[] = [
-    ...(dataset.dimensions || []).map((dimension) => {
-      const factTableColumn = dimension.factTableColumn;
-      const dimMetaEN = dimension.metadata?.find((meta) => meta.language.includes('en'));
-      const dimMetaCY = dimension.metadata?.find((meta) => meta.language.includes('cy'));
-      const dimNameEN = dimMetaEN?.name === factTableColumn ? '' : dimMetaEN?.name;
-      const dimNameCY = dimMetaCY?.name === factTableColumn ? '' : dimMetaCY?.name;
-
-      return {
-        type: 'dimension',
-        key: dimension.factTableColumn,
-        english: dimNameEN,
-        cymraeg: dimNameCY,
-        id: dimension.id
-      };
-    }),
-    ...metadataKeys.map((prop) => ({
-      type: 'metadata',
-      key: prop,
-      english: metadataEN?.[prop] as string,
-      cymraeg: metadataCY?.[prop] as string
-    })),
-    ...(revision.relatedLinks || []).map((link: RelatedLink) => ({
-      type: 'link',
-      key: link.id,
-      english: link.labelEN,
-      cymraeg: link.labelCY
-    }))
-  ];
-
-  return includeIds ? translations : translations.map((row) => pick(row, ['type', 'key', 'english', 'cymraeg']));
-};
 
 const parseUploadedTranslations = async (fileBuffer: Buffer): Promise<TranslationDTO[]> => {
   const translations: TranslationDTO[] = [];
@@ -111,6 +66,7 @@ translationRouter.get(
         action: 'export',
         entity: 'translations',
         entityId: revision.id,
+        data: { translations },
         userId: req.user?.id,
         client: 'sw3-frontend'
       });
